fix: add explicit casts to suppress MSVC C4244/C4267 conversion warnings (#4753)
Add static_cast to prevent implicit narrowing conversions that trigger MSVC warnings:

1. cb_explore_adf_large_action_space.cc: Cast float seed to uint64_t
2. hashstring.h: Cast size_t to uint32_t in hashstring return
3. action_score.h: Cast ptrdiff_t to difference_type in iterator subtraction
4. vw.h: Cast uint64_t namespace hash to uint32_t for hash_func_t calls

These casts make the intentional narrowing explicit and suppress warnings
while preserving existing behavior.

+++ vowpalwabbit/core/src/reductions/cb/cb_explore_adf_large_action_space.cc
@@ -302,7 +302,7 @@ std::shared_ptr<VW::LEARNER::learner> make_las_with_impl(VW::setup_base_i& stack
     bool apply_shrink_factor, size_t thread_pool_size, size_t block_size, size_t action_cache_slack,
     bool use_explicit_simd)
 {
-  float seed = (all.get_random_state()->get_random() + 1) * 10.f;
+  uint64_t seed = static_cast<uint64_t>((all.get_random_state()->get_random() + 1) * 10.f);
 
   auto data = VW::make_unique<cb_explore_adf_large_action_space<T, S>>(d, c, apply_shrink_factor, &all, seed,
       1 << all.initial_weights_config.num_bits, thread_pool_size, block_size, action_cache_slack, use_explicit_simd,
